requirements.txt
bottle
eventlet
bottle-swagger-2
psycopg2-binary
psycogreen
yoyo-migrations
